Moves GetSpaces into publically-available ParquetStack extensions.
