+++ ParquetClassLibrary/Parquets/ParquetStack.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Collections.Generic;
 using ParquetClassLibrary.Stubs;
 using ParquetClassLibrary.Utilities;
+using ParquetClassLibrary.Rooms;
 
 namespace ParquetClassLibrary.Parquets
 {
@@ -167,5 +169,28 @@ namespace ParquetClassLibrary.Parquets
             && in_position.Y > -1
             && in_position.X < in_subregion.GetLength(1)
             && in_position.Y < in_subregion.GetLength(0);
+
+        /// <summary>
+        /// Returns the set of <see cref="Space"/>s corresponding to the subregion.
+        /// </summary>
+        /// <param name="in_subregion">The collection of <see cref="ParquetStack"/>s to consider.</param>
+        /// <returns>The <see cref="Space"/>s defined by this subregion.</returns>
+        public static SpaceCollection GetSpaces(this ParquetStack[,] in_subregion)
+        {
+            var uniqueResults = new HashSet<Space();
+            var subregionRows = in_subregion.GetLength(0);
+            var subregionCols = in_subregion.GetLength(1);
+
+            for (var y = 0; y < subregionRows; y++)
+            {
+                for (var x = 0; x < subregionCols; x++)
+                {
+                    var currentSpace = new Space(x, y, in_subregion[y, x]);
+                    uniqueResults.Add(currentSpace);
+                }
+            }
+
+            return new SpaceCollection(uniqueResults);
+        }
     }
 }
